Spot: fix unit

+++ src/ltsmin-lib/ltl2spot.cpp
@@ -108,7 +108,7 @@ static int
 get_predicate_index(std::vector<std::string> pred_vec, std::string predicate) 
 {
   // iterate over the vector until the predicate matches
-  for (uint i=0; i<pred_vec.size(); i++) {
+  for (size_t i=0; i<pred_vec.size(); i++) {
     if (predicate.compare(pred_vec[i]) == 0)
       return i;
   }
@@ -138,7 +138,7 @@ create_ltsmin_buchi(spot::twa_graph_ptr& aut)
   uint32_t acceptance_set = 0;
   if (isTGBA) {
     HREassert (aut->num_sets() <= 32, "No more than 32 TGBA accepting sets supported.")
-    for (uint i=0; i<aut->num_sets(); i++)
+    for (size_t i=0; i<aut->num_sets(); i++)
       acceptance_set |= (1ULL << i);
   }
   else 
@@ -173,7 +173,7 @@ create_ltsmin_buchi(spot::twa_graph_ptr& aut)
     for (auto& t: aut->out(s)) {
       std::string cond = spot::bdd_format_formula(dict, t.cond);
       // count the number of '|' occurrences in the predicates
-      for (uint c_i=0; c_i<cond.length(); c_i++) {
+      for (size_t c_i=0; c_i<cond.length(); c_i++) {
         if (cond.at(c_i) == '|') {
           transition_count ++;
         }
@@ -214,7 +214,7 @@ create_ltsmin_buchi(spot::twa_graph_ptr& aut)
         std::string cond = spot::bdd_format_formula(dict, t.cond);
         int pred_start = -1;
         bool is_neg = false;
-        for (uint c_i=0; c_i<cond.length(); c_i++) {
+        for (size_t c_i=0; c_i<cond.length(); c_i++) {
           switch (cond.at(c_i)) {
             case '\"': {
               if (pred_start != -1) {
